requirements.txt
Flask==3.0.0
Pillow==10.1.0
exifread==3.0.0
folium==0.15.0
geopy==2.4.1
requests==2.31.0
gunicorn==21.2.0
numpy==1.24.3
flask-talisman==1.1.0
transformers==4.36.2
torch==2.1.2
google-cloud-vision==3.5.0
reverse-geocoder==1.5.1
